move common to root and delete scripts
