--- fabric/gradle.properties
@@ -1,1 +1,0 @@
-fabric.loom.disableProjectDependentMods = true
